+++ packages/search-form/index.ts
@@ -1,3 +1,10 @@
-import searchForm from "./src/search-form/index"
+import type { App } from "vue";
+import search from "./src/search-form/index.vue";
 
-export default [searchForm]
+export const SearchForm = Object.assign(search, {
+  install: function (app: App) {
+    app.component(search.name, search);
+  }
+});
+
+export default SearchForm;