+++ packages/search-form/src/search-form/index.js
@@ -1,5 +1,0 @@
-import searchForm from './index.vue';
-
-export default {
-    install: app => app.component(searchForm.name, searchForm)
-}