+++ packages/search-form/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tempoi/search-form
 
+## 1.1.1
+
+### Patch Changes
+
+- ...
+
 ## 2.1.0
 
 ### Minor Changes